DeepSec/Soulace/backend/requirements.txt
pytest
pytest-asyncio
httpx
pytest-mock
pytest-cov
python-dotenv
fastapi
uvicorn